+++ NEWS.md
@@ -6,7 +6,9 @@
 - added `tmap_mode_pool()` to restrict which modes are cycled by `ttm()` and `rtm()`, e.g. `library(tmap.mapgl); tmap_mode_pool(c("plot", "maplibre"))`
 - shiny dispatch methods added (to allow integration for other modes)
 - added support for terra's PackedSpatX objects 
+- added height to tmap_arrange (for interactive modes)
 - bug fixes (most importantly regarding >=1000 polygons in view mode)
+
 
 # tmap 4.3